Run browser tests locally
They error when run through Sauce Labs and it's hard to debug why.

The GitHub Actions runners include Chrome, Edge and Firefox along with
their WebDriver drivers, so the tests should work locally in CI too.

I've dropped testing for Internet Explorer and Edge because IE 11 is
end-of-life and Edge is based on Chromium these days so is effectively
covered by testing Chrome.

The Firefox and Chrome testing now also just runs against whatever is
installed for simplicity's sake.